requirements.txt
fastapi==0.111.1
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.21
pinecone-client==5.0.1
psycopg2==2.9.10
pydantic==2.8.2
python-dotenv==1.0.1
sentence-transformers==3.2.1
sqlalchemy===2.0.35
tqdm==4.66.4
uvicorn==0.30.3